Add Workouts endpoint

+++ GymProgresser.Application/Workouts/Validators/BaseWorkoutValidator.cs
@@ -12,7 +12,7 @@ namespace GymProgresser.Application.Workouts.Validators
     {
         public BaseWorkoutValidator()
         {
-            RuleFor(d => (int)d.GetType().GetProperty("DurationMin")!.GetValue(d)!)
+            RuleFor(d => (double)d.GetType().GetProperty("DurationMin")!.GetValue(d)!)
                 .GreaterThan(0)
                 .WithMessage("Trening musi trwać dłużej niż 0 minut");
         }
